move all logreg tests to sklearnex
